Added a way to start the puzzle and answer the puzzle based on the data loaded in
Need to add the functionality to all directions. Thinking of which class to put it in. Possibly the Puzzle class? I have solvingPuzzle as a inner game loop within my game driver class at the moment. Will come up with a plan and start to tidy up the code again.